+++ src/main/java/step3/dto/Car.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.dto;
 
 public class Car {
     private int entryNumber;
