+++ src/main/java/step3/dto/Competition.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.dto;
 
 public class Competition {
