+++ src/main/java/step3/view/ResultView.java
@@ -1,6 +1,6 @@
 package step3.view;
 
-import step3.Car;
+import step3.dto.Car;
 
 import java.util.List;
